Move tasks state to Pinia store (#6)
